ewrwe

+++ backend.card/LinkPara.Card.Application/Commons/Models/PaycoreModels/CardModels/AdditionalLimit.cs
@@ -3,7 +3,7 @@
     public class AdditionalLimit
     {
         public int CurrencyCode { get; set; }
-        public int CurrentLimit { get; set; }
+        public decimal CurrentLimit { get; set; }
         public int LimitRatio { get; set; }
         public string EffectType { get; set; }
         public string UsageType { get; set; }
